upd: alunoinfomodal pagination work // still bugs in table / aluno info visualization
